pyproject.toml
[tool.poetry]
name = "şañıraq-kz"
version = "0.1.0"
description = ""
authors = ["Bekarys <bekarys20050812@gmail.com>"]
readme = "README.md"
packages = [{include = "şañıraq"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.1"}
sqlalchemy = "^2.0.32"
psycopg2 = "^2.9.9"
pyjwt = "^2.9.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
